+++ core/interfaces.go
@@ -188,3 +188,8 @@ type Firebase interface {
 type Mailer interface {
 	SendMail(toEmail string, subject string, body string) error
 }
+
+// Core exposes Core APIs for the driver adapters
+type Core interface {
+	RetrieveCoreUserAccountByCriteria(accountCriteria map[string]interface{}, appID *string, orgID *string) ([]model.CoreAccount, error)
+}