+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Set logger [#80](https://github.com/rokwire/notifications-building-block/issues/80)
 - Add support of read/unread all user messages [#112](https://github.com/rokwire/notifications-building-block/issues/112)
+- Send notification by account data [#97](https://github.com/rokwire/notifications-building-block/issues/97)
 
 ### Changed
 - Use auth library authorization [#75](https://github.com/rokwire/notifications-building-block/issues/75)